added final setup for auth0

+++ backend/server/src/main/java/com/libbooks/server/controller/AdminController.java
@@ -3,7 +3,9 @@ package com.libbooks.server.controller;
 import com.libbooks.server.request.AddBookRequest;
 import com.libbooks.server.service.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,6 +15,9 @@ import java.util.List;
 @RequestMapping("/api/admin")
 public class AdminController {
 
+    @Value("${okta.oauth2.groupsClaim}")
+    private String groupsClaim;
+
     private AdminService adminService;
 
     @Autowired
@@ -21,46 +26,46 @@ public class AdminController {
     }
 
     @PutMapping("/secure/increase/book/quantity")
-    public void increaseBookQuantity(@RequestParam Long bookId) throws Exception { // @AuthenticationPrincipal Jwt jwt,
-//        List<String> roles = jwt.getClaimAsStringList("https://luv2code-react-library.com/roles");
-//        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
-//
-//        if (admin == null || !admin.equals("admin")) {
-//            throw new Exception("Administration page only");
-//        }
+    public void increaseBookQuantity(@AuthenticationPrincipal Jwt jwt, @RequestParam Long bookId) throws Exception {
+        List<String> roles = jwt.getClaimAsStringList(groupsClaim);
+        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
+
+        if (admin == null || !admin.equals("admin")) {
+            throw new Exception("Administration page only");
+        }
         adminService.increaseBookQuantity(bookId);
     }
 
     @PutMapping("/secure/decrease/book/quantity")
-    public void decreaseBookQuantity(@RequestParam Long bookId) throws Exception { // @AuthenticationPrincipal Jwt jwt,
-//        List<String> roles = jwt.getClaimAsStringList("https://luv2code-react-library.com/roles");
-//        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
-//
-//        if (admin == null || !admin.equals("admin")) {
-//            throw new Exception("Administration page only");
-//        }
+    public void decreaseBookQuantity(@AuthenticationPrincipal Jwt jwt, @RequestParam Long bookId) throws Exception {
+        List<String> roles = jwt.getClaimAsStringList(groupsClaim);
+        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
+
+        if (admin == null || !admin.equals("admin")) {
+            throw new Exception("Administration page only");
+        }
         adminService.decreaseBookQuantity(bookId);
     }
 
     @PostMapping("/secure/add/book")
-    public void postBook(@RequestBody AddBookRequest addBookRequest) throws Exception { // @AuthenticationPrincipal Jwt jwt,
-//        List<String> roles = jwt.getClaimAsStringList("https://luv2code-react-library.com/roles");
-//        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
-//
-//        if (admin == null || !admin.equals("admin")) {
-//            throw new Exception("Administration page only");
-//        }
+    public void postBook(@AuthenticationPrincipal Jwt jwt, @RequestBody AddBookRequest addBookRequest) throws Exception {
+        List<String> roles = jwt.getClaimAsStringList(groupsClaim);
+        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
+
+        if (admin == null || !admin.equals("admin")) {
+            throw new Exception("Administration page only");
+        }
         adminService.postBook(addBookRequest);
     }
 
     @DeleteMapping("/secure/delete/book")
-    public void deleteBook(@RequestParam Long bookId) throws Exception { // @AuthenticationPrincipal Jwt jwt,
-//        List<String> roles = jwt.getClaimAsStringList("https://luv2code-react-library.com/roles");
-//        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
-//
-//        if (admin == null || !admin.equals("admin")) {
-//            throw new Exception("Administration page only");
-//        }
+    public void deleteBook(@AuthenticationPrincipal Jwt jwt, @RequestParam Long bookId) throws Exception {
+        List<String> roles = jwt.getClaimAsStringList(groupsClaim);
+        String admin = roles != null && !roles.isEmpty() ? roles.get(0) : null;
+
+        if (admin == null || !admin.equals("admin")) {
+            throw new Exception("Administration page only");
+        }
         adminService.deleteBook(bookId);
     }
